Fix code insight deserialization (data points link generation)
